Dims and constant folding with type checking
