+++ example/exampleWidgetQt/symboltreeview.h
@@ -4,12 +4,12 @@
 #include <QTreeWidget>
 
 #include <hope_parse_tree.h>
+#include <hope_static_pass.h>
 #include <hope_symbol_table.h>
-#include <hope_type_resolver.h>
 
 class SymbolTreeView : public QTreeWidget {
 public:
-    SymbolTreeView(const Hope::Code::SymbolTable& symbol_table, const Hope::Code::TypeResolver& ts);
+    SymbolTreeView(const Hope::Code::SymbolTable& symbol_table, const Hope::Code::StaticPass& ts);
 };
 
 #endif // SYMBOLTREEVIEW_H
