+++ include/typeset_model.h
@@ -9,8 +9,7 @@
 #include "hope_scanner.h"
 #include "hope_parser.h"
 #include "hope_symbol_build_pass.h"
-#include "hope_type_resolver.h"
-#include "hope_optimiser.h"
+#include "hope_static_pass.h"
 #include "hope_interpreter.h"
 
 namespace Hope {
@@ -30,8 +29,7 @@ public:
     Code::Scanner scanner = Code::Scanner(this);
     Code::Parser parser = Code::Parser(scanner, this);
     Code::SymbolTableBuilder symbol_builder = Code::SymbolTableBuilder(parser.parse_tree, this);
-    Code::TypeResolver type_resolver = Code::TypeResolver(parser.parse_tree, symbol_builder.symbol_table, errors, warnings);
-    Code::Optimiser optimiser = Code::Optimiser(parser.parse_tree);
+    Code::StaticPass static_pass = Code::StaticPass(parser.parse_tree, symbol_builder.symbol_table, errors, warnings);
     Code::Interpreter interpreter;
     std::vector<Code::Error> errors;
     std::vector<Code::Error> warnings;
