+++ meta/parse_nodes.csv
@@ -199,3 +199,4 @@ LINEAR_SOLVE,A⁻¹b,NUMERIC,NUMERIC,,,
 INVERT,⁻¹,NUMERIC,,,,
 NORM_SQUARED,‖x‖²,NUMERIC,,,,
 DO_NOTHING,-DN-,,,,,
+REASSIGN_SUBSCRIPT,xₙ ← y,,,,,
