+++ src/typeset_model.cpp
@@ -539,8 +539,7 @@ void Model::performSemanticFormatting(){
     if(!is_output){
         parser.parseAll();
         symbol_builder.resolveSymbols();
-        type_resolver.resolve();
-        if(errors.empty()) optimiser.optimise();
+        static_pass.resolve();
     }
 }
