+++ test/CMakeLists.txt
@@ -121,8 +121,6 @@ set(SOURCES
     ${SRC}/hope_interpreter.cpp
     ${SRC}/hope_interpreter.h
     ${INCLUDE}/hope_logging.h
-    ${SRC}/hope_optimiser.cpp
-    ${SRC}/hope_optimiser.h
     ${SRC}/hope_parse_tree.cpp
     ${SRC}/hope_parse_tree.h
     ${SRC}/hope_parser.cpp
@@ -133,6 +131,8 @@ set(SOURCES
     ${INCLUDE}/hope_serial_unicode.h
     ${SRC}/hope_stack.cpp
     ${SRC}/hope_stack.h
+    ${SRC}/hope_static_pass.cpp
+    ${SRC}/hope_static_pass.h
     ${SRC}/hope_symbol_build_pass.cpp
     ${SRC}/hope_symbol_build_pass.h
     ${SRC}/hope_symbol_link_pass.cpp
@@ -140,8 +140,6 @@ set(SOURCES
     ${SRC}/hope_symbol_table.cpp
     ${SRC}/hope_symbol_table.h
     ${SRC}/hope_token.h
-    ${SRC}/hope_type_resolver.cpp
-    ${SRC}/hope_type_resolver.h
     ${SRC}/hope_unicode.h
     ${SRC}/hope_value.cpp
     ${SRC}/hope_value.h
